Resolve lethal temp and seasonality with the simulation (fixes #7)
Now we are tracking unresolved dates as well as step number.
This allows us to resolve season in the the loop excetuting the simulation
which is needed for the lethal temperature to be addressed there
which in turn allows the lethal temperature to be applied in the season as well.

The lethal temperature is now applied at most once a year.
Results are identical using SLF case with 6-11 season and lethal 1.