Rasters aggregated over runs are now doubles (#17)
Average, stddev, and probability in series and final are now handled
as floating point. Results are more precise and different from
the previous state which was an underestimate (for average up to 1/num_runs
different from the current value). A new object is created for each
computation which in an addition to several other objects already
being created in this context (before one of the inputs was reused).

This uses PoPS with raster-type-mixing updates and uses
the ncol, nrow constructors instead of stealing dimensions.
The numerical type typedefs are now used directly to use
whatever is in the definition.