Remove i.vi from the test setup
i.vi is changing behavior between GRASS GIS 7.8 and 8.0, outputting
float instead of double. Here, the creation of host map is very
sensitive, so the result of the test changes beween 7.6 (for what the
test was created) and 7.9 (future 8.0). Now computing NDVI manually
using r.mapcalc expression instead of i.vi. This changes the test
result, so the reference values were changed.

This was tested with code when the test was added (2f958d2) and
when more test were added (40b538ec) with both 7.6 (release branch)
and 7.9 (current master). The results are the same also for current
code (36a4ed1).

Also rounding the host map (convers to int) and using the dict syntax
for reference values.