Read null values as zeros (#21)
Read nulls as 0 for all inputs. Optional writing of nulls as zeros (not used).

The use of the (typesafe) strong enum is important to avoid any call or overload confusion
and it is useful to create a readble call.

A new test (copy of existing basic test) has nulls in input.