Add survival rate parameter (#51)
The addition of survival rate requires an additional parameter plus date settings. This adds it as a input raster series and generalizes code for reading raster names for raster series from a file which is common between survival rate and temperatures. Tests are for 100% and 0% survival rate.

Additionally, this adds number of steps as verbose message and fixes seed print (extra char d in output).